Release 11.2

+++ src/bp/ui-studio/src/web/views/Content/List/index.js
@@ -2,26 +2,23 @@ import React, { Component } from 'react'
 import classnames from 'classnames'
 import _ from 'lodash'
 import moment from 'moment'
-
-import { Checkbox, Table, Button, FormControl, FormGroup, Tooltip, OverlayTrigger, Glyphicon } from 'react-bootstrap'
-
-const style = require('./style.scss')
+import ReactTable from 'react-table'
+import 'react-table/react-table.css'
+import { Button, Glyphicon, FormGroup, FormControl } from 'react-bootstrap'
+import style from './style.scss'
 
 export default class ListView extends Component {
   state = {
     checkedIds: [],
     allChecked: false,
-    search: ''
+    pageSize: 10,
+    pages: 0,
+    filters: [],
+    sortOrder: []
   }
 
   componentDidMount() {
-    const fn = () => {
-      if (this.props.handleSearch) {
-        this.props.handleSearch(this.state.search)
-      }
-    }
-
-    this.debouncedHandleSearch = _.debounce(fn, 1000)
+    this.debouncedHandleSearch = _.debounce(() => this.launchSearch(), 1000)
   }
 
   handleCheckboxChanged(id) {
@@ -59,12 +56,14 @@ export default class ListView extends Component {
   }
 
   handleDeleteSelected = () => {
-    this.props.handleDeleteSelected(this.state.checkedIds)
+    if (window.confirm(`Do you really want to delete ${this.state.checkedIds.length} items?`)) {
+      this.props.handleDeleteSelected(this.state.checkedIds)
 
-    this.setState({
-      checkedIds: [],
-      allChecked: false
-    })
+      this.setState({
+        checkedIds: [],
+        allChecked: false
+      })
+    }
   }
 
   handleCloneSelected = () => {
@@ -78,92 +77,15 @@ export default class ListView extends Component {
 
   handleSearchChanged = event => {
     this.setState({
-      search: event.target.value
+      searchTerm: event.target.value
     })
 
     this.debouncedHandleSearch && this.debouncedHandleSearch()
   }
 
-  renderTableHeader() {
-    return (
-      <tr>
-        <th />
-        <th>Content Id</th>
-        <th>Content Type</th>
-        <th>Preview</th>
-        <th>Created on</th>
-      </tr>
-    )
-  }
-
-  renderContentItem = (m, i) => {
-    const handleEdit = () => !this.props.readOnly && this.props.handleEdit(m.id, m.contentType)
-    const checked = _.includes(this.state.checkedIds, m.id)
-    const className = classnames(style.item, { [style.selected]: checked })
-
-    return (
-      <tr className={className} key={i}>
-        {!this.props.readOnly && (
-          <td style={{ width: '2%', minWidth: '34px' }}>
-            <Checkbox checked={checked} onChange={() => this.handleCheckboxChanged(m.id, m.contentType)} />
-          </td>
-        )}
-        <td style={{ width: '16%' }}>{'#!' + m.id}</td>
-        <td style={{ width: '8%' }}>{m.contentType}</td>
-        <td style={{ width: '58%' }} onClick={handleEdit}>
-          {m.previewText}
-        </td>
-        <td style={{ width: '18%' }}>{moment(m.createdOn).format('MMMM Do YYYY, h:mm')}</td>
-        {!this.props.readOnly && (
-          <td>
-            <Button bsSize="small" onClick={handleEdit}>
-              <Glyphicon glyph="pencil" />
-            </Button>
-          </td>
-        )}
-      </tr>
-    )
-  }
-
-  renderTable() {
-    if (this.props.contentItems && this.props.contentItems.length === 0) {
-      const message = this.props.readOnly
-        ? "There's no content here."
-        : "There's no content yet. You can create some using the 'Add' button."
-      return <div className={style.empty}>{message}</div>
-    }
-
-    return (
-      <div className={style.container}>
-        <Table striped bordered condensed hover>
-          <tbody>{this.props.contentItems.map(this.renderContentItem)}</tbody>
-        </Table>
-      </div>
-    )
-  }
-
-  renderPaging() {
-    const count = this.props.count
-
-    let from = (this.props.page - 1) * this.props.itemsPerPage + 1
-    let to = this.props.page * this.props.itemsPerPage
-
-    from = count !== 0 ? from : 0
-    to = to <= count ? to : count
-
-    const text = `${from} - ${to} of ${count}`
-
-    return <span className={style.paging}>{text}</span>
-  }
-
   renderActionButtons() {
     return (
       <span>
-        {!this.props.readOnly && (
-          <Button onClick={this.handleAllCheckedChanged}>
-            <Checkbox checked={this.state.allChecked} onChange={this.handleAllCheckedChanged} />
-          </Button>
-        )}
         <Button onClick={this.props.handleRefresh} title="Refresh">
           <i className="material-icons">refresh</i>
         </Button>
@@ -185,19 +107,15 @@ export default class ListView extends Component {
     )
   }
 
-  renderPagingButtons() {
+  renderHeader() {
+    const leftCls = classnames('pull-left', style.left)
     return (
-      <span className={style.pagingButtons}>
-        <Button onClick={this.props.handlePrevious} disabled={this.props.page === 1}>
-          <i className="material-icons">keyboard_arrow_left</i>
-        </Button>
-        <Button
-          onClick={this.props.handleNext}
-          disabled={this.props.page * this.props.itemsPerPage >= this.props.count}
-        >
-          <i className="material-icons">keyboard_arrow_right</i>
-        </Button>
-      </span>
+      <div className={style.header}>
+        <div className={leftCls}>
+          {this.renderActionButtons()}
+          {this.renderSearchBar()}
+        </div>
+      </div>
     )
   }
 
@@ -209,21 +127,165 @@ export default class ListView extends Component {
     )
   }
 
-  renderHeader() {
-    const leftCls = classnames('pull-left', style.left)
-    const rightCls = classnames('pull-right', style.right)
+  launchSearch = () => {
+    const searchQuery = {
+      from: this.state.page * this.state.pageSize,
+      count: this.state.pageSize,
+      sortOrder: this.state.sortOrder,
+      filters: this.state.filters,
+      searchTerm: this.state.searchTerm
+    }
+
+    this.props.handleSearch(searchQuery)
+  }
+
+  fetchData = state => {
+    const filters = state.filtered.map(filter => {
+      return { column: filter.id, value: filter.value }
+    })
+    const sortOrder = state.sorted.map(sort => {
+      return { column: sort.id, desc: sort.desc }
+    })
+    const hasTextChanged = !_.isEqual(this.state.filters, filters)
+
+    this.setState(
+      {
+        pageSize: state.pageSize,
+        page: state.page,
+        sortOrder,
+        filters
+      },
+      () => {
+        hasTextChanged ? this.debouncedHandleSearch && this.debouncedHandleSearch() : this.launchSearch()
+      }
+    )
+  }
+
+  renderFilterPlaceholder = placeholder => ({ filter, onChange }) => (
+    <input
+      type="text"
+      placeholder={placeholder}
+      value={filter ? filter.value : ''}
+      style={{ width: '100%' }}
+      onChange={event => onChange(event.target.value)}
+    />
+  )
+
+  onRowClick = (state, rowInfo, column, instance) => {
+    return {
+      onClick: (e, handleOriginal) => {
+        if (column.id !== 'checkbox' && !this.props.readOnly) {
+          const { id, contentType } = rowInfo.original
+          this.props.handleEdit(id, contentType)
+        }
+
+        if (handleOriginal) {
+          handleOriginal()
+        }
+      }
+    }
+  }
+
+  getTableColumns() {
+    return [
+      {
+        Header: x => {
+          return (
+            <input
+              type="checkbox"
+              className="checkbox"
+              checked={this.state.allChecked === 1}
+              ref={input => {
+                if (input) {
+                  input.indeterminate = this.state.allChecked === 2
+                }
+              }}
+              onChange={() => this.handleAllCheckedChanged()}
+            />
+          )
+        },
+        Cell: ({ original }) => {
+          const checked = _.includes(this.state.checkedIds, original.id)
+          return (
+            <input
+              type="checkbox"
+              className="checkbox"
+              checked={checked}
+              onChange={() => this.handleCheckboxChanged(original.id)}
+            />
+          )
+        },
+        id: 'checkbox',
+        accessor: '',
+        filterable: false,
+        sortable: false,
+        width: 35
+      },
+      {
+        Header: 'ID',
+        Cell: x => `#!${x.value}`,
+        Filter: this.renderFilterPlaceholder('Filter'),
+        accessor: 'id',
+        width: 170
+      },
+      {
+        Header: 'Content Type',
+        Filter: this.renderFilterPlaceholder('Filter'),
+        accessor: 'contentType',
+        width: 150
+      },
+      {
+        Header: 'Preview',
+        Filter: this.renderFilterPlaceholder('Filter'),
+        accessor: 'previewText'
+      },
+      {
+        Header: 'Modified On',
+        Cell: x => (x.original.modifiedOn ? moment(x.original.modifiedOn).format('MMM Do YYYY, h:mm') : 'Never'),
+        accessor: 'modifiedOn',
+        filterable: false,
+        width: 150
+      },
+      {
+        Header: 'Created On',
+        Cell: x => (x.original.createdOn ? moment(x.original.createdOn).format('MMM Do YYYY, h:mm') : 'Never'),
+        accessor: 'createdOn',
+        filterable: false,
+        width: 150
+      },
+      {
+        Cell: !this.props.readOnly && (
+          <Button bsSize="xs">
+            <Glyphicon glyph="pencil" />
+          </Button>
+        ),
+        filterable: false,
+        width: 45
+      }
+    ]
+  }
+
+  renderTable() {
+    const pageCount = Math.ceil(this.props.count / this.state.pageSize)
+    const noDataMessage = this.props.readOnly
+      ? "There's no content here."
+      : "There's no content yet. You can create some using the 'Add' button."
 
     return (
-      <div className={style.header}>
-        <div className={leftCls}>
-          {this.renderActionButtons()}
-          {this.renderSearchBar()}
-        </div>
-        <div className={rightCls}>
-          {this.renderPaging()}
-          {this.renderPagingButtons()}
-        </div>
-      </div>
+      <ReactTable
+        columns={this.getTableColumns()}
+        data={this.props.contentItems}
+        onFetchData={this.fetchData}
+        onPageSizeChange={(pageSize, page) => this.setState({ page, pageSize })}
+        getTdProps={this.onRowClick}
+        defaultPageSize={this.state.pageSize}
+        defaultSorted={[{ id: 'modifiedOn', desc: true }]}
+        noDataText={noDataMessage}
+        className="-striped -highlight"
+        pages={pageCount}
+        manual
+        filterable
+      />
     )
   }
